Mise à jour des GenericListItemService et Dao pour refléter le type de l'id des GenericListItems (Long).
git-svn-id: svn+ssh://svn.projects.openwide.fr/svnroot/owsi-tools/trunk/owsi-core@1757 d3474844-eb8e-4abc-b058-2b321fed648b

+++ owsi-core/owsi-core-components/owsi-core-component-jpa-more/src/main/java/fr/openwide/core/jpa/more/business/generic/dao/GenericListItemDaoImpl.java
@@ -40,12 +40,12 @@ public class GenericListItemDaoImpl extends JpaDaoSupport implements IGenericLis
 	}
 	
 	@Override
-	public <E extends GenericListItem<?>> E getEntity(Class<E> clazz, Integer id) {
+	public <E extends GenericListItem<?>> E getEntity(Class<E> clazz, Long id) {
 		return super.getEntity(clazz, id);
 	}
 	
 	@Override
-	public <E extends GenericListItem<?>> E getById(Class<E> clazz, Integer id) {
+	public <E extends GenericListItem<?>> E getById(Class<E> clazz, Long id) {
 		return super.getEntity(clazz, id);
 	}
